+++ src/edit/resources.ts
@@ -1,19 +1,15 @@
 import EventEmitter from 'eventemitter3';
 import SceneManager from './sceneManager/sceneManager';
-import { LoaderManager } from './loader';
-import { LoaderResourceType } from '@/edit/loader';
-export interface assetType {
-  name: string;
-  url: string;
-  type: LoaderResourceType;
-}
+import { ILoaderResource, LoaderManager } from './loader';
+
+export interface Asset extends ILoaderResource {};
 class Resources extends EventEmitter {
   public loader: LoaderManager;
   private toLoad: number;
   private loaded: number;
   public progress: number;
   public items: { [key: string]: any };
-  constructor(assets: assetType[]) {
+  constructor(assets: Asset[]) {
     super();
     this.loader = SceneManager.loader;
     this.loaded = 0;
@@ -25,18 +21,30 @@ class Resources extends EventEmitter {
     this.loadAssets(assets);
   }
 
-  loadAssets(assets: assetType[]) {
+  loadAssets(assets: Asset[]) {
     this.loaded = 0;
     this.toLoad = assets.length;
     assets.forEach(async (asset) => {
       const item = await this.loader.load(asset);
       this.loaded++;
-      this.items[asset.name] = item;
+      this.items[asset.url] = item;
       this.progress = this.loaded / this.toLoad;
       if (this.loaded === this.toLoad) {
         this.emit('loaded');
       }
     });
+  }
+
+  async loadAsset(asset: Asset) {
+    if(this.items[asset.url]) {
+      return this.items[asset.url];
+    }
+    
+    const item = await this.loader.load(asset);
+    if(item) {
+      this.items[asset.url] = item;
+    }
+    return item;
   }
 
   destroy() {}